[minor] factor module registration into its own file
