+++ src/mig/modules/modules.go
@@ -50,24 +50,6 @@ type Result struct {
 	Errors        []string    `json:"errors"`
 }
 
-// Stores details about the registration of a module
-type Registration struct {
-	Runner func() interface{}
-}
-
-// Available stores a list of activated module with their registration
-var Available = make(map[string]Registration)
-
-// Register adds a module to the list of available modules
-func Register(name string, runner func() interface{}) {
-	if _, exist := Available[name]; exist {
-		panic("Register: a module named " + name + " has already been registered.\nAre you trying to import the same module twice?")
-	}
-	newmodule := &Registration{}
-	newmodule.Runner = runner
-	Available[name] = *newmodule
-}
-
 // Moduler provides the interface to a Module
 type Moduler interface {
 	Run(io.Reader) string
